+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `cli`: Add skeleton for plugin output
 - `cli/sync`: Add support for alias "-"
 - `cli/move`: Implement move command
+- `cli/rm`: Add verbose option
 
 ### Changed